+++ src/test/java/com/wireguard/api/peer/controller/UpdatePeerControllerTest.java
@@ -1,0 +1,7 @@
+package com.wireguard.api.peer.controller;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class UpdatePeerControllerTest {
+
+}
